+++ ome_zarr/writer.py
@@ -3,8 +3,10 @@
 """
 import logging
 import warnings
+from pathlib import Path
 from typing import Any, Dict, List, Optional, Tuple, Union
 
+import dask.array as da
 import numpy as np
 import zarr
 
@@ -15,10 +17,13 @@ from .types import JSONDict
 
 LOGGER = logging.getLogger("ome_zarr.writer")
 
+ListOfArrayLike = Union[List[da.Array], List[np.ndarray]]
+ArrayLike = Union[da.Array, np.ndarray]
+
 
 def _get_valid_axes(
-    ndim: int = None,
-    axes: Union[str, List[str], List[Dict[str, str]]] = None,
+    ndim: Optional[int] = None,
+    axes: Optional[Union[str, List[str], List[Dict[str, str]]]] = None,
     fmt: Format = CurrentFormat(),
 ) -> Union[None, List[str], List[Dict[str, str]]]:
     """Returns list of axes valid for fmt.version or raise exception if invalid"""
@@ -32,10 +37,10 @@ def _get_valid_axes(
     if axes is None:
         if ndim == 2:
             axes = ["y", "x"]
-            LOGGER.info("Auto using axes %s for 2D data" % axes)
+            LOGGER.info("Auto using axes %s for 2D data", axes)
         elif ndim == 5:
             axes = ["t", "c", "z", "y", "x"]
-            LOGGER.info("Auto using axes %s for 5D data" % axes)
+            LOGGER.info("Auto using axes %s for 5D data", axes)
         else:
             raise ValueError(
                 "axes must be provided. Can't be guessed for 3D or 4D data"
@@ -59,7 +64,6 @@ def _get_valid_axes(
 def _validate_well_images(
     images: List[Union[str, dict]], fmt: Format = CurrentFormat()
 ) -> List[dict]:
-
     VALID_KEYS = [
         "acquisition",
         "path",
@@ -70,7 +74,7 @@ def _validate_well_images(
             validated_images.append({"path": str(image)})
         elif isinstance(image, dict):
             if any(e not in VALID_KEYS for e in image.keys()):
-                LOGGER.debug("f{image} contains unspecified keys")
+                LOGGER.debug("%s contains unspecified keys", image)
             if "path" not in image:
                 raise ValueError(f"{image} must contain a path key")
             if not isinstance(image["path"], str):
@@ -86,7 +90,6 @@ def _validate_well_images(
 def _validate_plate_acquisitions(
     acquisitions: List[Dict], fmt: Format = CurrentFormat()
 ) -> List[Dict]:
-
     VALID_KEYS = [
         "id",
         "name",
@@ -100,7 +103,7 @@ def _validate_plate_acquisitions(
         if not isinstance(acquisition, dict):
             raise ValueError(f"{acquisition} must be a dictionary")
         if any(e not in VALID_KEYS for e in acquisition.keys()):
-            LOGGER.debug("f{acquisition} contains unspecified keys")
+            LOGGER.debug("%s contains unspecified keys", acquisition)
         if "id" not in acquisition:
             raise ValueError(f"{acquisition} must contain an id key")
         if not isinstance(acquisition["id"], int):
@@ -112,7 +115,6 @@ def _validate_plate_rows_columns(
     rows_or_columns: List[str],
     fmt: Format = CurrentFormat(),
 ) -> List[dict]:
-
     if len(set(rows_or_columns)) != len(rows_or_columns):
         raise ValueError(f"{rows_or_columns} must contain unique elements")
     validated_list = []
@@ -126,7 +128,6 @@ def _validate_plate_rows_columns(
 def _validate_datasets(
     datasets: List[dict], dims: int, fmt: Format = CurrentFormat()
 ) -> List[Dict]:
-
     if datasets is None or len(datasets) == 0:
         raise ValueError("Empty datasets list")
     transformations = []
@@ -151,7 +152,6 @@ def _validate_plate_wells(
     columns: List[str],
     fmt: Format = CurrentFormat(),
 ) -> List[dict]:
-
     validated_wells = []
     if wells is None or len(wells) == 0:
         raise ValueError("Empty wells list")
@@ -169,19 +169,20 @@ def _validate_plate_wells(
 
 
 def write_multiscale(
-    pyramid: List,
+    pyramid: ListOfArrayLike,
     group: zarr.Group,
-    chunks: Union[Tuple[Any, ...], int] = None,
+    chunks: Optional[Union[Tuple[Any, ...], int]] = None,
     fmt: Format = CurrentFormat(),
-    axes: Union[str, List[str], List[Dict[str, str]]] = None,
-    coordinate_transformations: List[List[Dict[str, Any]]] = None,
-    storage_options: Union[JSONDict, List[JSONDict]] = None,
+    axes: Optional[Union[str, List[str], List[Dict[str, str]]]] = None,
+    coordinate_transformations: Optional[List[List[Dict[str, Any]]]] = None,
+    storage_options: Optional[Union[JSONDict, List[JSONDict]]] = None,
+    name: Optional[str] = None,
     **metadata: Union[str, JSONDict, List[JSONDict]],
 ) -> None:
     """
     Write a pyramid with multiscale metadata to disk.
 
-    :type pyramid: list of :class:`numpy.ndarray`
+    :type pyramid: list of :class:`numpy.ndarray` or :class:`dask.array.Array`
     :param pyramid:
         The image data to save. Largest level first. All image arrays MUST be up to
         5-dimensional with dimensions ordered (t, c, z, y, x)
@@ -214,7 +215,6 @@ def write_multiscale(
         One can provide different chunk size for each level of a pyramid using this
         option.
     """
-
     dims = len(pyramid[0].shape)
     axes = _get_valid_axes(dims, axes, fmt)
 
@@ -225,13 +225,8 @@ Please use the 'storage_options' argument instead."""
 
     datasets: List[dict] = []
     for path, data in enumerate(pyramid):
-        options = {}
-        if storage_options:
-            options = (
-                storage_options
-                if not isinstance(storage_options, list)
-                else storage_options[path]
-            )
+        options = _resolve_storage_options(storage_options, path)
+
         # ensure that the chunk dimensions match the image dimensions
         # (which might have been changed for versions 0.1 or 0.2)
         # if chunks are explicitly set in the storage options
@@ -240,7 +235,22 @@ Please use the 'storage_options' argument instead."""
         # chunks_opt = options.pop("chunks", None)
         if chunks_opt is not None:
             chunks_opt = _retuple(chunks_opt, data.shape)
-        group.create_dataset(str(path), data=data, chunks=chunks_opt, **options)
+
+        if isinstance(data, da.Array):
+            if chunks_opt is not None:
+                data = da.array(data).rechunk(chunks=chunks_opt)
+                options["chunks"] = chunks_opt
+            da.to_zarr(
+                arr=data,
+                url=group.store,
+                component=str(Path(group.path, str(path))),
+                storage_options=options,
+                compressor=options.get("compressor", zarr.storage.default_compressor),
+                dimension_separator=group._store._dimension_separator,
+            )
+        else:
+            group.create_dataset(str(path), data=data, chunks=chunks_opt, **options)
+
         datasets.append({"path": str(path)})
 
     if coordinate_transformations is None:
@@ -255,14 +265,15 @@ Please use the 'storage_options' argument instead."""
         for dataset, transform in zip(datasets, coordinate_transformations):
             dataset["coordinateTransformations"] = transform
 
-    write_multiscales_metadata(group, datasets, fmt, axes, **metadata)
+    write_multiscales_metadata(group, datasets, fmt, axes, name, **metadata)
 
 
 def write_multiscales_metadata(
     group: zarr.Group,
     datasets: List[dict],
     fmt: Format = CurrentFormat(),
-    axes: Union[str, List[str], List[Dict[str, str]]] = None,
+    axes: Optional[Union[str, List[str], List[Dict[str, str]]]] = None,
+    name: Optional[str] = None,
     **metadata: Union[str, JSONDict, List[JSONDict]],
 ) -> None:
     """
@@ -302,6 +313,7 @@ def write_multiscales_metadata(
         dict(
             version=fmt.version,
             datasets=_validate_datasets(datasets, ndim, fmt),
+            name=name if name else group.name,
             **metadata,
         )
     ]
@@ -317,9 +329,9 @@ def write_plate_metadata(
     columns: List[str],
     wells: List[Union[str, dict]],
     fmt: Format = CurrentFormat(),
-    acquisitions: List[dict] = None,
-    field_count: int = None,
-    name: str = None,
+    acquisitions: Optional[List[dict]] = None,
+    field_count: Optional[int] = None,
+    name: Optional[str] = None,
 ) -> None:
     """
     Write the plate metadata in the group.
@@ -385,24 +397,24 @@ def write_well_metadata(
 
 
 def write_image(
-    image: np.ndarray,
+    image: ArrayLike,
     group: zarr.Group,
     scaler: Scaler = Scaler(),
-    chunks: Union[Tuple[Any, ...], int] = None,
+    chunks: Optional[Union[Tuple[Any, ...], int]] = None,
     fmt: Format = CurrentFormat(),
-    axes: Union[str, List[str], List[Dict[str, str]]] = None,
-    coordinate_transformations: List[List[Dict[str, Any]]] = None,
-    storage_options: Union[JSONDict, List[JSONDict]] = None,
+    axes: Optional[Union[str, List[str], List[Dict[str, str]]]] = None,
+    coordinate_transformations: Optional[List[List[Dict[str, Any]]]] = None,
+    storage_options: Optional[Union[JSONDict, List[JSONDict]]] = None,
     **metadata: Union[str, JSONDict, List[JSONDict]],
 ) -> None:
     """Writes an image to the zarr store according to ome-zarr specification
 
-    :type image: :class:`numpy.ndarray`
+    :type image: :class:`numpy.ndarray` or `dask.array.Array`
     :param image:
       The image data to save. A downsampling of the data will be computed
       if the scaler argument is non-None.
       Image array MUST be up to 5-dimensional with dimensions
-      ordered (t, c, z, y, x)
+      ordered (t, c, z, y, x).  Image can be a numpy or dask Array.
     :type group: :class:`zarr.hierarchy.Group`
     :param group: The group within the zarr store to write the metadata in.
     :type scaler: :class:`ome_zarr.scale.Scaler`
@@ -435,24 +447,140 @@ def write_image(
         One can provide different chunk size for each level of a pyramid using this
         option.
     """
-    mip, axes = _create_mip(image, fmt, scaler, axes)
-    write_multiscale(
-        mip,
-        group,
-        chunks=chunks,
-        fmt=fmt,
-        axes=axes,
-        coordinate_transformations=coordinate_transformations,
-        storage_options=storage_options,
-        **metadata,
+    if isinstance(image, da.Array):
+        _write_dask_image(
+            image,
+            group,
+            scaler,
+            chunks=chunks,
+            fmt=fmt,
+            axes=axes,
+            coordinate_transformations=coordinate_transformations,
+            storage_options=storage_options,
+            name=None,
+            **metadata,
+        )
+    else:
+        mip, axes = _create_mip(image, fmt, scaler, axes)
+        write_multiscale(
+            mip,
+            group,
+            chunks=chunks,
+            fmt=fmt,
+            axes=axes,
+            coordinate_transformations=coordinate_transformations,
+            storage_options=storage_options,
+            name=None,
+            **metadata,
+        )
+
+
+def _resolve_storage_options(
+    storage_options: Union[JSONDict, List[JSONDict], None], path: int
+) -> JSONDict:
+    options = {}
+    if storage_options:
+        options = (
+            storage_options.copy()
+            if not isinstance(storage_options, list)
+            else storage_options[path]
+        )
+    return options
+
+
+def _write_dask_image(
+    image: da.Array,
+    group: zarr.Group,
+    scaler: Scaler = Scaler(),
+    chunks: Optional[Union[Tuple[Any, ...], int]] = None,
+    fmt: Format = CurrentFormat(),
+    axes: Optional[Union[str, List[str], List[Dict[str, str]]]] = None,
+    coordinate_transformations: Optional[List[List[Dict[str, Any]]]] = None,
+    storage_options: Optional[Union[JSONDict, List[JSONDict]]] = None,
+    name: Optional[str] = None,
+    **metadata: Union[str, JSONDict, List[JSONDict]],
+) -> None:
+    if fmt.version in ("0.1", "0.2"):
+        # v0.1 and v0.2 are strictly 5D
+        shape_5d: Tuple[Any, ...] = (*(1,) * (5 - image.ndim), *image.shape)
+        image = image.reshape(shape_5d)
+        # and we don't need axes
+        axes = None
+
+    dims = len(image.shape)
+    axes = _get_valid_axes(dims, axes, fmt)
+
+    if chunks is not None:
+        msg = """The 'chunks' argument is deprecated and will be removed in version 0.5.
+Please use the 'storage_options' argument instead."""
+        warnings.warn(msg, DeprecationWarning)
+
+    datasets: List[dict] = []
+    delayed = []
+
+    # for path, data in enumerate(pyramid):
+    max_layer: int = scaler.max_layer if scaler is not None else 0
+    shapes = []
+    for path in range(0, max_layer + 1):
+        # LOGGER.debug(f"write_image path: {path}")
+        options = _resolve_storage_options(storage_options, path)
+
+        # don't downsample top level of pyramid
+        if str(path) != "0" and scaler is not None:
+            image = scaler.resize_image(image)
+
+        # ensure that the chunk dimensions match the image dimensions
+        # (which might have been changed for versions 0.1 or 0.2)
+        # if chunks are explicitly set in the storage options
+        chunks_opt = options.pop("chunks", chunks)
+        # switch to this code in 0.5
+        # chunks_opt = options.pop("chunks", None)
+        if chunks_opt is not None:
+            chunks_opt = _retuple(chunks_opt, image.shape)
+            image = da.array(image).rechunk(chunks=chunks_opt)
+            options["chunks"] = chunks_opt
+        LOGGER.debug("chunks_opt: %s", chunks_opt)
+        shapes.append(image.shape)
+
+        LOGGER.debug(
+            "write dask.array to_zarr shape: %s, dtype: %s", image.shape, image.dtype
+        )
+        delayed.append(
+            da.to_zarr(
+                arr=image,
+                url=group.store,
+                component=str(Path(group.path, str(path))),
+                storage_options=options,
+                compute=False,
+                compressor=options.get("compressor", zarr.storage.default_compressor),
+                dimension_separator=group._store._dimension_separator,
+            )
+        )
+        datasets.append({"path": str(path)})
+
+    da.compute(*delayed)
+
+    if coordinate_transformations is None:
+        # shapes = [data.shape for data in delayed]
+        coordinate_transformations = fmt.generate_coordinate_transformations(shapes)
+
+    # we validate again later, but this catches length mismatch before zip(datasets...)
+    fmt.validate_coordinate_transformations(
+        dims, len(datasets), coordinate_transformations
     )
+    if coordinate_transformations is not None:
+        for dataset, transform in zip(datasets, coordinate_transformations):
+            dataset["coordinateTransformations"] = transform
+
+    write_multiscales_metadata(group, datasets, fmt, axes, name, **metadata)
 
 
 def write_label_metadata(
     group: zarr.Group,
     name: str,
-    colors: List[JSONDict] = None,
-    properties: List[JSONDict] = None,
+    colors: Optional[List[JSONDict]] = None,
+    properties: Optional[List[JSONDict]] = None,
+    fmt: Format = CurrentFormat(),
     **metadata: Union[List[JSONDict], JSONDict, str],
 ) -> None:
     """
@@ -476,6 +604,10 @@ def write_label_metadata(
       Each dict specifies additional properties for one label.
       It must contain the field "label-value"
       and may contain arbitrary additional properties.
+    :type fmt: :class:`ome_zarr.format.Format`, optional
+    :param fmt:
+      The format of the ome_zarr data which should be used.
+      Defaults to the most current.
     """
     label_group = group[name]
     image_label_metadata = {**metadata}
@@ -483,6 +615,7 @@ def write_label_metadata(
         image_label_metadata["colors"] = colors
     if properties is not None:
         image_label_metadata["properties"] = properties
+    image_label_metadata["version"] = fmt.version
     label_group.attrs["image-label"] = image_label_metadata
 
     label_list = group.attrs.get("labels", [])
@@ -494,12 +627,12 @@ def write_multiscale_labels(
     pyramid: List,
     group: zarr.Group,
     name: str,
-    chunks: Union[Tuple[Any, ...], int] = None,
+    chunks: Optional[Union[Tuple[Any, ...], int]] = None,
     fmt: Format = CurrentFormat(),
-    axes: Union[str, List[str], List[Dict[str, str]]] = None,
-    coordinate_transformations: List[List[Dict[str, Any]]] = None,
-    storage_options: Union[JSONDict, List[JSONDict]] = None,
-    label_metadata: JSONDict = None,
+    axes: Optional[Union[str, List[str], List[Dict[str, str]]]] = None,
+    coordinate_transformations: Optional[List[List[Dict[str, Any]]]] = None,
+    storage_options: Optional[Union[JSONDict, List[JSONDict]]] = None,
+    label_metadata: Optional[JSONDict] = None,
     **metadata: JSONDict,
 ) -> None:
     """
@@ -559,21 +692,24 @@ def write_multiscale_labels(
         **metadata,
     )
     write_label_metadata(
-        group["labels"], name, **({} if label_metadata is None else label_metadata)
+        group["labels"],
+        name,
+        fmt=fmt,
+        **({} if label_metadata is None else label_metadata),
     )
 
 
 def write_labels(
-    labels: np.ndarray,
+    labels: Union[np.ndarray, da.Array],
     group: zarr.Group,
     name: str,
     scaler: Scaler = Scaler(),
-    chunks: Union[Tuple[Any, ...], int] = None,
+    chunks: Optional[Union[Tuple[Any, ...], int]] = None,
     fmt: Format = CurrentFormat(),
-    axes: Union[str, List[str], List[Dict[str, str]]] = None,
-    coordinate_transformations: List[List[Dict[str, Any]]] = None,
-    storage_options: Union[JSONDict, List[JSONDict]] = None,
-    label_metadata: JSONDict = None,
+    axes: Optional[Union[str, List[str], List[Dict[str, str]]]] = None,
+    coordinate_transformations: Optional[List[List[Dict[str, Any]]]] = None,
+    storage_options: Optional[Union[JSONDict, List[JSONDict]]] = None,
+    label_metadata: Optional[JSONDict] = None,
     **metadata: JSONDict,
 ) -> None:
     """
@@ -625,18 +761,38 @@ def write_labels(
     :param label_metadata:
       Image label metadata. See :meth:`write_label_metadata` for details
     """
-    mip, axes = _create_mip(labels, fmt, scaler, axes)
-    write_multiscale_labels(
-        mip,
-        group,
+    sub_group = group.require_group(f"labels/{name}")
+    if isinstance(labels, da.Array):
+        _write_dask_image(
+            labels,
+            sub_group,
+            scaler,
+            chunks=chunks,
+            fmt=fmt,
+            axes=axes,
+            coordinate_transformations=coordinate_transformations,
+            storage_options=storage_options,
+            name=name,
+            **metadata,
+        )
+    else:
+        mip, axes = _create_mip(labels, fmt, scaler, axes)
+        write_multiscale(
+            mip,
+            sub_group,
+            chunks=chunks,
+            fmt=fmt,
+            axes=axes,
+            coordinate_transformations=coordinate_transformations,
+            storage_options=storage_options,
+            name=name,
+            **metadata,
+        )
+    write_label_metadata(
+        group=group["labels"],
         name=name,
-        chunks=chunks,
         fmt=fmt,
-        axes=axes,
-        coordinate_transformations=coordinate_transformations,
-        storage_options=storage_options,
-        label_metadata=label_metadata,
-        **metadata,
+        **({} if label_metadata is None else label_metadata),
     )
 
 
@@ -675,6 +831,12 @@ def _create_mip(
 def _retuple(
     chunks: Union[Tuple[Any, ...], int], shape: Tuple[Any, ...]
 ) -> Tuple[Any, ...]:
+    """
+    Expand chunks to match shape.
+
+    E.g. if chunks is (64, 64) and shape is (3, 4, 5, 1028, 1028)
+    return (3, 4, 5, 64, 64)
+    """
 
     _chunks: Tuple[Any, ...]
     if isinstance(chunks, int):